% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPCC.R
\name{classify_climate}
\alias{classify_climate}
\title{Assign local climate to IPCC category}
\usage{
classify_climate(mat, maxmt, map, pet, elev, frost)
}
\arguments{
\item{mat}{Mean annual temperature (degrees C).}

\item{maxmt}{Maximum mean monthly temperature (degrees C).}

\item{map}{Mean annual precipitation (mm).}

\item{pet}{Mean potential evapotranspiration (mm).}

\item{elev}{Elevation (m).}

\item{frost}{Mean frost days per year.}
}
\value{
The IPPC climate zone based on MAT, MAP, PET, frost, and elevation.
}
\description{
Assign local climate to IPCC category
}
