% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPCC.R
\name{classify_soil_WRB}
\alias{classify_soil_WRB}
\title{Classification scheme for mineral soil types based on World Reference Base
for Soil Resources (WRB) classification.}
\usage{
classify_soil_WRB(sand, clay, wrb_class)
}
\arguments{
\item{sand}{sand content (\%).}

\item{clay}{clay content (\%).}

\item{wrb_class}{World Reference Base for Soil Resources (WRB) classification}
}
\value{
The IPPC soil type.
}
\description{
Classification scheme for mineral soil types based on World Reference Base
for Soil Resources (WRB) classification.
}
