% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{save_raster}
\alias{save_raster}
\title{Save a raster to file.}
\usage{
save_raster(x, path = ".", fname = NULL, type = "tif", overwrite = NULL, ...)
}
\arguments{
\item{x}{raster* object to save.}

\item{path}{path in which to save.  May be relative or absolute.}

\item{fname}{character vector, length one.
Base of file name (not including the extension, which is determined by file type).
If no file name is provided, default behaviour is to use the object name as the file name.}

\item{type}{file type to save.  Default is ".tif".  Accepts any type handled by raster::writeRaster.}

\item{overwrite}{logical. Whether to overwrite existing files of same name. Default
value of NULL causes function to ask for permission if an existing file would be overwritten.}

\item{...}{optional parameters passed through to raster::writeRaster.}
}
\value{
invisibly returns the full name (including path) of the saved raster.
}
\description{
saves raster object to file. By default, asks permission to overwrite an existing file.
}
