% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{which.level}
\alias{which.level}
\title{Find ID in raster attribute table}
\usage{
which.level(r, val, id.col = 1, val.col = 2)
}
\arguments{
\item{r}{raster* object - must have an associated raster attribute table (RAT).}

\item{val}{value of the attribute to lookup.}

\item{id.col}{column in the RAT containing the ID to lookup.}

\item{val.col}{column in the RAT containing the attribute to match.}
}
\value{
Returns the ID value corresponding to the attribute value
}
\description{
Given a raster and attribute value, finds the corresponding raster value
for that attribute.  Essentially a "vlookup" for raster attribute tables.
}
