% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{wmax}
\alias{wmax}
\title{Variant of which.max}
\usage{
wmax(x)
}
\arguments{
\item{x}{vector.}
}
\value{
The ordinal position of the maximum element in x. Unlike `which.max`,
wmax returns NA if no element of x is positive definite.
}
\description{
helper function to find which raster layer from a stack contains data
}
